Extract gem updater helpers into cli utility module
