+++ lib/polisher/cli/all.rb
@@ -10,11 +10,14 @@ require 'polisher/cli/sources'
 require 'polisher/cli/targets'
 require 'polisher/cli/specifier'
 require 'polisher/cli/format'
+require 'polisher/cli/fedora'
 require 'polisher/cli/conf'
+require 'polisher/cli/paths'
 require 'polisher/cli/status'
 require 'polisher/cli/deps'
 require 'polisher/cli/profiles'
 require 'polisher/cli/cross_ref'
+require 'polisher/cli/updater'
 
 module Polisher
   module CLI
